requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-openai
openai
neo4j
pydantic
pydantic-settings
python-dotenv
httpx
requests
beautifulsoup4
lxml
pypdf
tiktoken
cachetools
redis
orjson
tenacity
pytest
pytest-asyncio
faiss-cpu
numpy
huggingface_hub
